Moved config into celery_config file and added celery_serde
